Move utils tests to dedicated file
